Make repositories created with new Chorus load in old versions
* Move the allownumberbranch into the fixutf8 extensions.
  (adding a new extension breaks backward compatibility)
* Adjust installer source
* Remove the allownumberbranch as an extension
