+++ src/LibChorusTests/VcsDrivers/Mercurial/MercurialTestUtilities.cs
@@ -34,15 +34,9 @@ namespace LibChorus.Tests.VcsDrivers.Mercurial
 #endif
 			extensions.Add("hgext.graphlog", ""); //for more easily readable diagnostic logs
 			extensions.Add("convert", ""); //for catastrophic repair in case of repo corruption
-#if !MONO
 			string fixUtfFolder = FileLocator.GetDirectoryDistributedWithApplication(false, "MercurialExtensions", "fixutf8");
 			if (!string.IsNullOrEmpty(fixUtfFolder))
 				extensions.Add("fixutf8", Path.Combine(fixUtfFolder, "fixutf8.py"));
-#endif
-			// Add extension to allow creation of number only branches
-			var allownumberext = FileLocator.GetDirectoryDistributedWithApplication(false, "MercurialExtensions", "allownumberbranch");
-			if(!string.IsNullOrEmpty(allownumberext))
-				extensions.Add("allownumberbranch", Path.Combine(allownumberext, "allownumberbranch.py"));
 
 			var doc = HgRepository.GetMercurialConfigInMercurialFolder();
 			SetExtensions(doc, extensions);
